+++ modules/siamux.go
@@ -1,9 +1,11 @@
 package modules
 
 import (
+	"errors"
 	"os"
 	"path/filepath"
 
+	"gitlab.com/NebulousLabs/Sia/build"
 	"gitlab.com/NebulousLabs/Sia/crypto"
 	"gitlab.com/NebulousLabs/Sia/persist"
 	"gitlab.com/NebulousLabs/Sia/types"
@@ -18,6 +20,13 @@ const (
 
 // NewSiaMux returns a new SiaMux object
 func NewSiaMux(persistDir, address string) (*siamux.SiaMux, error) {
+	// can't use relative path
+	if !filepath.IsAbs(persistDir) {
+		err := errors.New("siamux path needs to be absolute")
+		build.Critical(err)
+		return nil, err
+	}
+
 	// ensure the persist directory exists
 	err := os.MkdirAll(persistDir, 0700)
 	if err != nil {